code reorganization and stream option
